+++ src/app/_component/modal/toDoCard/type.ts
@@ -1,0 +1,16 @@
+export interface TodoProps {
+  mainTitle: string;
+}
+
+export interface ToDoCardDetailProps {
+  columnId: number;
+  id: number;
+  createdAt: string;
+  updatedAt: string;
+  title: string;
+  description: string;
+  tags: string[];
+  dueDate: string;
+  imageUrl: string;
+  assignee: { profileImageUrl: string; nickname: string; id: number };
+}
